Split create connection component
